refactor(eventservice/test): Move mock schema store to dedicated test helper
- Centralizes the `mockSchemaStore` definition and its constructor.
- Improves code organization by separating common test utilities into `test_helper.go`.
- Enables easier reuse of the mock across different test files and packages.
- Updates all call sites to use the exported `NewMockSchemaStore`.
