+++ pkg/eventservice/event_service_test.go
@@ -16,8 +16,6 @@ package eventservice
 import (
 	"context"
 	"fmt"
-	"math"
-	"sort"
 	"sync"
 	"testing"
 	"time"
@@ -36,7 +34,6 @@ import (
 	"github.com/pingcap/ticdc/pkg/messaging"
 	"github.com/pingcap/ticdc/pkg/node"
 	"github.com/pingcap/ticdc/pkg/pdutil"
-	"github.com/pingcap/tidb/pkg/kv"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
 )
@@ -45,7 +42,7 @@ func startEventService(
 	ctx context.Context, t *testing.T,
 	mc messaging.MessageCenter, mockStore eventstore.EventStore,
 ) *eventService {
-	mockSchemaStore := newMockSchemaStore()
+	mockSchemaStore := NewMockSchemaStore()
 	mockPDClock := pdutil.NewClock4Test()
 	appcontext.SetService(appcontext.DefaultPDClock, mockPDClock)
 	appcontext.SetService(appcontext.MessageCenter, mc)
@@ -373,134 +370,6 @@ func (m *mockEventIterator) Close() (int64, error) {
 }
 
 var _ schemastore.SchemaStore = &mockSchemaStore{}
-
-type mockSchemaStore struct {
-	DDLEvents map[common.TableID][]commonEvent.DDLEvent
-	TableInfo map[common.TableID]*mockVersionTableInfo
-
-	resolvedTs     uint64
-	maxDDLCommitTs uint64
-}
-
-type mockVersionTableInfo struct {
-	tableInfos    []*common.TableInfo
-	deleteVersion uint64
-}
-
-func newMockSchemaStore() *mockSchemaStore {
-	return &mockSchemaStore{
-		DDLEvents:      make(map[common.TableID][]commonEvent.DDLEvent),
-		TableInfo:      make(map[common.TableID]*mockVersionTableInfo),
-		resolvedTs:     math.MaxUint64,
-		maxDDLCommitTs: math.MaxUint64,
-	}
-}
-
-func (m *mockSchemaStore) Name() string {
-	return "mockSchemaStore"
-}
-
-func (m *mockSchemaStore) Run(ctx context.Context) error {
-	return nil
-}
-
-func (m *mockSchemaStore) Close(ctx context.Context) error {
-	return nil
-}
-
-func (m *mockSchemaStore) DeleteTable(id common.TableID, ts common.Ts) {
-	if info, ok := m.TableInfo[id]; ok {
-		info.deleteVersion = uint64(ts)
-	} else {
-		m.TableInfo[id] = &mockVersionTableInfo{
-			tableInfos:    make([]*common.TableInfo, 0),
-			deleteVersion: uint64(ts),
-		}
-	}
-}
-
-func (m *mockSchemaStore) AppendDDLEvent(id common.TableID, ddls ...commonEvent.DDLEvent) {
-	for _, ddl := range ddls {
-		m.DDLEvents[id] = append(m.DDLEvents[id], ddl)
-		if _, ok := m.TableInfo[id]; !ok {
-			m.TableInfo[id] = &mockVersionTableInfo{
-				tableInfos:    make([]*common.TableInfo, 0),
-				deleteVersion: math.MaxUint64,
-			}
-		}
-		m.TableInfo[id].tableInfos = append(m.TableInfo[id].tableInfos, ddl.TableInfo)
-	}
-}
-
-func (m *mockSchemaStore) GetTableInfo(keyspaceID uint32, tableID common.TableID, ts common.Ts) (*common.TableInfo, error) {
-	if info, ok := m.TableInfo[tableID]; ok {
-		if info.deleteVersion <= uint64(ts) {
-			return nil, &schemastore.TableDeletedError{}
-		}
-		infos := m.TableInfo[tableID].tableInfos
-		idx := sort.Search(len(infos), func(i int) bool {
-			return infos[i].GetUpdateTS() > uint64(ts)
-		})
-		if idx == 0 {
-			return nil, nil
-		}
-		return infos[idx-1], nil
-	}
-	return nil, nil
-}
-
-func (m *mockSchemaStore) GetAllPhysicalTables(keyspaceID uint32, snapTs uint64, filter filter.Filter) ([]commonEvent.Table, error) {
-	return nil, nil
-}
-
-func (m *mockSchemaStore) GetTableDDLEventState(keyspaceID uint32, tableID int64) (schemastore.DDLEventState, error) {
-	return schemastore.DDLEventState{
-		ResolvedTs:       m.resolvedTs,
-		MaxEventCommitTs: m.maxDDLCommitTs,
-	}, nil
-}
-
-func (m *mockSchemaStore) RegisterTable(
-	keyspaceID uint32,
-	tableID int64,
-	startTS common.Ts,
-) error {
-	return nil
-}
-
-func (m *mockSchemaStore) UnregisterTable(_ uint32, _ int64) error {
-	return nil
-}
-
-// GetNextDDLEvents returns the next ddl event which finishedTs is within the range (start, end]
-func (m *mockSchemaStore) FetchTableDDLEvents(keyspaceID uint32, dispatcherID common.DispatcherID, tableID int64, tableFilter filter.Filter, start, end uint64) ([]commonEvent.DDLEvent, error) {
-	events := m.DDLEvents[tableID]
-	if len(events) == 0 {
-		return nil, nil
-	}
-	l := sort.Search(len(events), func(i int) bool {
-		return events[i].FinishedTs > start
-	})
-	if l == len(events) {
-		return nil, nil
-	}
-	r := sort.Search(len(events), func(i int) bool {
-		return events[i].FinishedTs > end
-	})
-	return events[l:r], nil
-}
-
-func (m *mockSchemaStore) FetchTableTriggerDDLEvents(keyspaceID uint32, tableFilter filter.Filter, start uint64, limit int) ([]commonEvent.DDLEvent, uint64, error) {
-	return nil, 0, nil
-}
-
-func (m *mockSchemaStore) RegisterKeyspace(ctx context.Context, keyspaceName string) error {
-	return nil
-}
-
-func (m *mockSchemaStore) GetKVStorage(keyspaceID uint32) (kv.Storage, error) {
-	return nil, nil
-}
 
 type mockSpanStats struct {
 	mu                 sync.RWMutex
